+++ crates/compiler/src/songs.rs
@@ -17,8 +17,8 @@ use crate::driver_constants::{
 };
 use crate::envelope::Envelope;
 use crate::errors::{SongError, SongTooLargeError, ValueError};
-use crate::mml::VibratoState;
 use crate::mml::{MetaData, MmlInstrument, Section};
+use crate::mml::{SlurredNoteState, VibratoState};
 use crate::notes::Note;
 use crate::sound_effects::CompiledSoundEffect;
 use crate::time::{TickClock, TickCounter, TickCounterWithLoopFlag};
@@ -76,6 +76,7 @@ pub struct Subroutine {
     pub last_instrument: Option<usize>,
     pub last_envelope: Option<Envelope>,
     pub vibrato: VibratoState,
+    pub prev_slurred_note: SlurredNoteState,
     pub changes_song_tempo: bool,
 }